chore(plugin): extract C ABI types into winisland-plugin-api crate
