+++ Cargo.toml
@@ -3,6 +3,9 @@ name = "WinIsland"
 version = "1.0.0"
 edition = "2024"
 
+[workspace]
+members = ["crates/winisland-plugin-api"]
+
 [dependencies]
 cpal = "0.15"
 realfft = "3.3"
@@ -25,6 +28,7 @@ rfd = "0.17.2"
 once_cell = "1.19"
 libloading = "0.9"
 log = "0.4"
+winisland-plugin-api = { path = "crates/winisland-plugin-api" }
 
 [build-dependencies]
 winres = "0.1"
